use CCPiTestClass as base class
