+++ Wrappers/Python/test/testclass.py
@@ -16,19 +16,17 @@
 #   limitations under the License.
 
 import unittest
-from cil.framework import ImageGeometry, ImageData, BlockDataContainer, DataContainer
-import numpy
+from cil.framework import DataContainer
+import numpy as np
 
 def dt(steps):
     return steps[-1] - steps[-2]
 
 class CCPiTestClass(unittest.TestCase):
     def assertBlockDataContainerEqual(self, container1, container2):
-        print ("assert Block Data Container Equal")
         self.assertTrue(issubclass(container1.__class__, container2.__class__))
         for col in range(container1.shape[0]):
             if issubclass(container1.get_item(col).__class__, DataContainer):
-                print ("Checking col ", col)
                 self.assertNumpyArrayEqual(
                     container1.get_item(col).as_array(), 
                     container2.get_item(col).as_array()
@@ -36,23 +34,24 @@ class CCPiTestClass(unittest.TestCase):
             else:
                 self.assertBlockDataContainerEqual(container1.get_item(col),container2.get_item(col))
     
+    
+    def assertBlockDataContainerAlmostEqual(self, container1, container2, decimal=7):
+        self.assertTrue(issubclass(container1.__class__, container2.__class__))
+        for col in range(container1.shape[0]):
+            if issubclass(container1.get_item(col).__class__, DataContainer):
+                self.assertNumpyArrayAlmostEqual(
+                    container1.get_item(col).as_array(), 
+                    container2.get_item(col).as_array(), 
+                    decimal=decimal
+                    )
+            else:
+                self.assertBlockDataContainerAlmostEqual(container1.get_item(col),container2.get_item(col), decimal=decimal)
+
+
     def assertNumpyArrayEqual(self, first, second):
-        res = True
-        try:
-            numpy.testing.assert_array_equal(first, second)
-        except AssertionError as err:
-            res = False
-            print(err)
-        self.assertTrue(res)
+        np.testing.assert_array_equal(first, second)
+        
 
     def assertNumpyArrayAlmostEqual(self, first, second, decimal=6):
-        res = True
-        try:
-            numpy.testing.assert_array_almost_equal(first, second, decimal)
-        except AssertionError as err:
-            res = False
-            print(err)
-            print("expected " , second)
-            print("actual " , first)
-
-        self.assertTrue(res)
+        np.testing.assert_array_almost_equal(first, second, decimal)
+
